tests: railscript: reorganize code into "helpers" file
